move v1 specific components out of /shared and into /v1
