+++ src/components/shared/formItems/index.tsx
@@ -4,7 +4,7 @@ import ProjectDescription from './ProjectDescription'
 import ProjectDiscord from './ProjectDiscord'
 import ProjectDiscountRate from './ProjectDiscountRate'
 import ProjectDuration from './ProjectDuration'
-import ProjectHandleFormItem from './ProjectHandle/ProjectHandleFormItem'
+import ProjectHandleFormItem from '../../v1/shared/formItems/ProjectHandle/ProjectHandleFormItem'
 import ProjectLink from './ProjectLink'
 import ProjectLogoUri from './ProjectLogoUri'
 import ProjectName from './ProjectName'
@@ -12,9 +12,9 @@ import ProjectPayButton from './ProjectPayButton'
 import ProjectPayDisclosure from './ProjectPayDisclosure'
 import ProjectReserved from './ProjectReserved'
 import ProjectTarget from './ProjectTarget'
-import ProjectTicketMods from './ProjectTicketMods'
+import ProjectTicketMods from '../../v1/ProjectTicketMods'
 import ProjectTwitter from './ProjectTwitter'
-import TokenRefs from './TokenRefs'
+import TokenRefs from '../../v1/shared/formItems/TokenRefs'
 import { ItemNoInput } from './ItemNoInput'
 
 export const FormItems = {
